Added concept of Stamina Stat.
I created a concept of PlayerStamina.  This is currently only consumed when the player sprints.  But it would be cool to have a means of being consumed by some sort of damage component.

IE:  IHealthDamagable implements IDamagable and IStaminaDamagable implements IDamagable.  Then the IdamagableProvider can provide the appropriate Idamagable.  Perhaps by using a SO enum.