Refactored Stats, Refactored Damage System, Created Minimap
I refactored the stat system so there are no "Concrete" implementations.  There are either StatComponent or RegenStatComponent. and they hold a reference to a SO to determine the StatType.  This allows DamageDealer to apply damage to any stat.

I also added a minimap.

See _Docs/TODO.txt for more info on how to refactor even more for more extensability.