Refactored DamageEffects to be simply Effects
Refactored DamageEffects to be simply Effects. This way they can easliy be plugged into the Interaction system too.  This is useful for playing oneshot audio and possibly a little camera shake when player warps.
I extended out the geometry and altered some of the previous geometry to better reflect my idea of how the gameplay should be used.
-Created a shader for the damage wall.
- Added some post processing effects and 2D lights